feat: 유저 및 임시여행 엔티티 컬럼 수정
